Corrected bug when selecting protoeme file

+++ TriFusion.py
@@ -3416,20 +3416,22 @@ class TriFusionApp(App):
         # Changes concerning the files tab
         if parent_obj == self.root.ids.file_sl:
 
-            # When button is normal (unselected) remove from active list
-            if value.state == "normal":
-                self.active_file_list.remove(self.filename_map[value.id])
-                self.active_alignment_list.remove_file(
-                    [self.filename_map[value.id]])
-            # When button is down (selected) add to active list
-            elif value.state == "down":
-                self.active_file_list.append(self.filename_map[value.id])
-                self.active_alignment_list.add_alignment(
-                    self.alignment_list.retrieve_alignment(
-                        self.filename_map[value.id]))
+            if self.filename_map[value.id] in self.active_file_list:
 
-            # Update label
-            self.update_file_label()
+                # When button is normal (unselected) remove from active list
+                if value.state == "normal":
+                    self.active_file_list.remove(self.filename_map[value.id])
+                    self.active_alignment_list.remove_file(
+                        [self.filename_map[value.id]])
+                # When button is down (selected) add to active list
+                elif value.state == "down":
+                    self.active_file_list.append(self.filename_map[value.id])
+                    self.active_alignment_list.add_alignment(
+                        self.alignment_list.retrieve_alignment(
+                            self.filename_map[value.id]))
+
+                # Update label
+                self.update_file_label()
 
         # Changes concerning the taxa tab
         if parent_obj == self.root.ids.taxa_sl:
@@ -6210,7 +6212,7 @@ class TriFusionApp(App):
                 opipe.filter_fasta(self.protein_min_len, self.protein_max_stop)
 
             if nm.k:
-                nm.t = "Running USearch"
+                nm.t = "Running USearch. This may take a while..."
                 nm.c = 4
                 opipe.allvsall_usearch("goodProteins.fasta",
                                   self.usearch_evalue,
